% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxlab1.R
\docType{package}
\name{boxlab1}
\alias{boxlab1}
\alias{boxlab1-package}
\title{boxlab1: euclidean and dijkstra algorithm}
\description{
conduct euclidean and dikstra algorithm for given input
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\author{
Maintainer: Boxi Zhang bossyzhang@gmail.com
}
