% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxlab1.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra algorithm.}
\usage{
dijkstra(cost, v)
}
\arguments{
\item{cost}{the data frame.}

\item{v}{the start point.}
}
\description{
This function uses dijkstra algorithm to find the shortest connection.
}
\examples{
dijkstra()
d <- dijkstra(cost=wiki_graph, v=1)
}
\references{
adapted from https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm .
}
