% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian algorithm.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{the first number.}

\item{y}{the second number.}
}
\description{
This function uses euclidian algorithm to find the greatest common divisor.
adapted from https://en.wikipedia.org/wiki/Euclidean_algorithm .
}
\examples{
euclidean()
d <- euclidean(x=100, y=1000)
}
