% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
wiki_graph
}
\description{
wiki_graph
\itemize{
  \item v1 numeric vector
  \item v2 numeric vector
  \item w numeric vector
}
}
\references{
\url{https://en.wikipedia.org/wiki/Graph}
}
\keyword{datasets}
